Revert "set INFO appender as stdout"
This reverts commit 1ac818b9d1c29801d671a2f7efcbf72347406fff.

This change was originally done somewhat haphazardly to get more info
in running sstableloader in some situations, and did not consider side
effects. Namely that is makes all tools that use (parts of) the origin
framework spew loads and loads of info to stdout.

This breaks such tools as sstabledump etc, which sort of rely on
a sanitized stdout for operation.

If we want more output from sstableloader, lets put the output there
and not modify logging for the whole universe.

Message-Id: <20181016113147.15494-1-calle@scylladb.com>