Add procps dependency
nodetool uses `free` indirectly through cassandra-env.sh so it will
not work if procps is not installed. This is the kind of tool that is
present everywhere, but we just found a user using "Debian9 Slim" in
which it is not.

It is a good practice to list this explicitly, then.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20181022190204.23091-1-glauber@scylladb.com>